refactor(desktop): finalize ChatPane split — orchestrator becomes index (Phase 5/5)
- UserTurn.tsx
- ConversationTurns.tsx (re-exported from index for external consumers)
- ChatPane.tsx → ChatPane/index.tsx; sibling imports normalized

External imports of `@/components/panes/ChatPane` resolve to
ChatPane/index.tsx unchanged. No runtime behavior delta.

Test file ChatPane.test.mjs source path updated to read from
ChatPane/index.tsx. Many assertions now fail because their
patterns live in sibling files; this is expected — the test scope
is unchanged from before this PR.
