+++ desktop/src/components/panes/ChatPane/index.tsx
@@ -113,7 +113,7 @@ import {
   type WorkspaceFileEntry,
 } from "@/lib/workspaceFiles";
 import { useWorkspaceSelection } from "@/lib/workspaceSelection";
-import * as modelCatalog from "../../../shared/model-catalog.js";
+import * as modelCatalog from "../../../../shared/model-catalog.js";
 import {
   type ChatAttachment,
   type ChatPaneVariant,
@@ -140,7 +140,7 @@ import {
   type ChatComposerMentionItem,
   type StreamTelemetryEntry,
   type ArtifactBrowserFilter,
-} from "./ChatPane/types";
+} from "./types";
 import {
   MAIN_SESSION_EVENT_BATCH_HEADER,
   BACKGROUND_DELIVERY_RETRY_STATUS_MESSAGE,
@@ -172,7 +172,7 @@ import {
   CHAT_MODEL_PRESETS,
   RUNTIME_MODEL_CAPABILITY_ALIASES,
   MENTION_TOKEN_PATTERN,
-} from "./ChatPane/constants";
+} from "./constants";
 import {
   initialBrowserState,
   attachmentLooksLikeImage,
@@ -194,16 +194,16 @@ import {
   inputIdFromHistoryMessage,
   historyMessagesInDisplayOrder,
   turnInputIdsFromHistoryMessages,
-} from "./ChatPane/helpers";
+} from "./helpers";
 import {
   HistoryRestoreSkeleton,
   IntegrationErrorBanner,
-} from "./ChatPane/skeletons";
+} from "./skeletons";
 import {
   AttachmentList,
   formatAttachmentSize,
-} from "./ChatPane/AttachmentList";
-import { ImageAttachmentPreviewModal } from "./ChatPane/ImageAttachmentPreviewModal";
+} from "./AttachmentList";
+import { ImageAttachmentPreviewModal } from "./ImageAttachmentPreviewModal";
 import {
   ArtifactBrowserModal,
   OutputArtifactIcon,
@@ -216,11 +216,13 @@ import {
   sortOutputs,
   sortOutputsLatestFirst,
   dedupeOutputsForDisplay,
-} from "./ChatPane/ArtifactBrowserModal";
-import { ChatHeader } from "./ChatPane/ChatHeader";
-import { QueuedSessionInputRail } from "./ChatPane/QueuedSessionInputRail";
-import { AssistantTurn } from "./ChatPane/AssistantTurn";
-import { Composer } from "./ChatPane/Composer";
+} from "./ArtifactBrowserModal";
+import { ChatHeader } from "./ChatHeader";
+import { QueuedSessionInputRail } from "./QueuedSessionInputRail";
+import { AssistantTurn } from "./AssistantTurn";
+import { Composer } from "./Composer";
+import { UserTurn } from "./UserTurn";
+import { ConversationTurns } from "./ConversationTurns";
 
 export type {
   ChatAssistantSegment,
@@ -235,6 +237,8 @@ export {
   ArtifactBrowserModal,
   sortOutputs,
   AssistantTurn,
+  UserTurn,
+  ConversationTurns,
 };
 
 function sessionUserId(
@@ -2379,34 +2383,6 @@ function latestVisibleChatMessageId(messages: ChatMessage[]): string {
     }
   }
   return "";
-}
-
-async function copyTextToClipboard(value: string) {
-  if (typeof navigator !== "undefined" && navigator.clipboard?.writeText) {
-    await navigator.clipboard.writeText(value);
-    return;
-  }
-  if (typeof document === "undefined") {
-    throw new Error("Clipboard is unavailable.");
-  }
-
-  const textarea = document.createElement("textarea");
-  textarea.value = value;
-  textarea.setAttribute("readonly", "");
-  textarea.style.position = "fixed";
-  textarea.style.opacity = "0";
-  textarea.style.pointerEvents = "none";
-  document.body.appendChild(textarea);
-  textarea.focus();
-  textarea.select();
-  textarea.setSelectionRange(0, textarea.value.length);
-
-  const copied = document.execCommand("copy");
-  document.body.removeChild(textarea);
-
-  if (!copied) {
-    throw new Error("Clipboard copy failed.");
-  }
 }
 
 function hasActiveChatSelection(container: HTMLDivElement | null) {
@@ -7487,349 +7463,6 @@ export function ChatPane({
         </div>
       </div>
     </PaneCard>
-  );
-}
-
-export const UserTurn = memo(UserTurnComponent, (prev, next) =>
-  prev.text === next.text &&
-  prev.createdAt === next.createdAt &&
-  prev.attachments === next.attachments,
-);
-
-function UserTurnComponent({
-  text,
-  createdAt,
-  attachments,
-  onPreviewAttachment,
-  onLinkClick,
-  onLocalLinkClick,
-}: {
-  text: string;
-  createdAt?: string;
-  attachments: ChatAttachment[];
-  onPreviewAttachment?: (attachment: AttachmentListItem) => void;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-}) {
-  const [copyFeedbackVisible, setCopyFeedbackVisible] = useState(false);
-  const copyResetTimerRef = useRef<number | null>(null);
-  const timeLabel = chatMessageTimeLabel(createdAt);
-  const canCopy = text.trim().length > 0;
-  const showHoverFooter = canCopy || Boolean(timeLabel);
-  const parsedQuotedSkills = useMemo(
-    () => parseSerializedQuotedSkillPrompt(text),
-    [text],
-  );
-  const userBubbleText = parsedQuotedSkills.body || text.trim();
-
-  const bubbleContentRef = useRef<HTMLDivElement>(null);
-  const [isExpanded, setIsExpanded] = useState(false);
-  const [showExpandButton, setShowExpandButton] = useState(false);
-
-  useEffect(() => {
-    const node = bubbleContentRef.current;
-    if (!node) {
-      return;
-    }
-    // 180px ~= 6–7 lines of chat-user-markdown at 0.875rem / 1.6 leading.
-    setShowExpandButton(node.scrollHeight > 188);
-  }, [userBubbleText]);
-
-  useEffect(() => {
-    return () => {
-      if (copyResetTimerRef.current !== null) {
-        window.clearTimeout(copyResetTimerRef.current);
-      }
-    };
-  }, []);
-
-  const handleCopy = async () => {
-    if (!canCopy) {
-      return;
-    }
-
-    try {
-      await copyTextToClipboard(text);
-    } catch {
-      return;
-    }
-    setCopyFeedbackVisible(true);
-    if (copyResetTimerRef.current !== null) {
-      window.clearTimeout(copyResetTimerRef.current);
-    }
-    copyResetTimerRef.current = window.setTimeout(() => {
-      setCopyFeedbackVisible(false);
-      copyResetTimerRef.current = null;
-    }, 1600);
-  };
-
-  return (
-    <div className="group/user-turn flex min-w-0 justify-end">
-      <div
-        className={`relative z-0 flex min-w-0 max-w-[80%] flex-col items-end gap-2 group-hover/user-turn:z-10 group-focus-within/user-turn:z-10 ${showHoverFooter ? "pb-7" : ""}`.trim()}
-      >
-        {parsedQuotedSkills.skillIds.length > 0 ? (
-          <div className="flex max-w-full flex-wrap justify-end gap-1">
-            {parsedQuotedSkills.skillIds.map((skillId) => (
-              <EntityChip
-                key={skillId}
-                icon={<Sparkles className="text-muted-foreground" />}
-                label={`/${skillId}`}
-              />
-            ))}
-          </div>
-        ) : null}
-        {userBubbleText ? (
-          <div className="theme-chat-user-bubble inline-flex min-w-0 max-w-full flex-col items-stretch rounded-lg px-3 py-1.5 text-foreground">
-            <div
-              ref={bubbleContentRef}
-              className="relative overflow-hidden transition-[max-height] duration-300 ease-out"
-              style={{
-                maxHeight: showExpandButton && !isExpanded ? 180 : undefined,
-              }}
-            >
-              <SimpleMarkdown
-                className="chat-markdown chat-user-markdown max-w-full"
-                onLinkClick={onLinkClick}
-                onLocalLinkClick={onLocalLinkClick}
-                renderMention={(handle) => (
-                  <EntityMention label={handle} />
-                )}
-              >
-                {injectMentionLinks(userBubbleText)}
-              </SimpleMarkdown>
-              {showExpandButton && !isExpanded ? (
-                <div
-                  aria-hidden
-                  className="pointer-events-none absolute inset-x-0 bottom-0 h-10"
-                  style={{
-                    background:
-                      "linear-gradient(to bottom, transparent, color-mix(in oklch, var(--muted) 85%, var(--foreground) 4%))",
-                  }}
-                />
-              ) : null}
-            </div>
-            {showExpandButton ? (
-              <button
-                type="button"
-                onClick={() => setIsExpanded((value) => !value)}
-                className="mt-1.5 self-start text-xs font-medium text-muted-foreground transition-colors hover:text-foreground"
-              >
-                {isExpanded ? "Show less" : "Show more"}
-              </button>
-            ) : null}
-          </div>
-        ) : null}
-        {attachments.length > 0 ? (
-          <AttachmentList
-            attachments={attachments}
-            className="justify-end"
-            onPreview={onPreviewAttachment}
-          />
-        ) : null}
-        {showHoverFooter ? (
-          <div className="absolute bottom-0 right-1 flex w-max min-w-max max-w-none items-center gap-2 whitespace-nowrap text-xs text-muted-foreground opacity-0 pointer-events-none transition-opacity duration-150 group-hover/user-turn:opacity-100 group-hover/user-turn:pointer-events-auto group-focus-within/user-turn:opacity-100 group-focus-within/user-turn:pointer-events-auto">
-            {canCopy ? (
-              <Button
-                type="button"
-                variant="ghost"
-                size="icon-xs"
-                aria-label={
-                  copyFeedbackVisible
-                    ? "Copied user message"
-                    : "Copy user message"
-                }
-                onClick={() => {
-                  void handleCopy();
-                }}
-                className="size-6 rounded-lg text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-              >
-                {copyFeedbackVisible ? (
-                  <Check className="size-3.5" strokeWidth={1.9} />
-                ) : (
-                  <Copy className="size-3.5" strokeWidth={1.9} />
-                )}
-              </Button>
-            ) : null}
-            {timeLabel ? (
-              <span className="select-none whitespace-nowrap tabular-nums">
-                {timeLabel}
-              </span>
-            ) : null}
-          </div>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-export function ConversationTurns<Message extends ChatMessage>({
-  messages,
-  assistantLabel,
-  assistantMode,
-  showExecutionInternals,
-  assistantFitToContent = false,
-  /** Drives the agent avatar's seed so each workspace has its own
-   *  persistent face. */
-  workspaceId,
-  onPreviewAttachment,
-  onOpenOutput,
-  onOpenAllArtifacts,
-  collapsedTraceByStepId,
-  onToggleTraceStep,
-  onLinkClick,
-  onLocalLinkClick,
-  memoryProposalAction,
-  editingMemoryProposalId,
-  memoryProposalDrafts,
-  onEditMemoryProposal,
-  onMemoryProposalDraftChange,
-  onAcceptMemoryProposal,
-  onDismissMemoryProposal,
-  assistantFooterAccessoryMessageId = null,
-  assistantFooterAccessory = null,
-  getMessageWrapperClassName,
-  liveAssistantTurn = null,
-}: {
-  messages: Message[];
-  assistantLabel: string;
-  assistantMode: string;
-  showExecutionInternals: boolean;
-  assistantFitToContent?: boolean;
-  workspaceId?: string | null;
-  onPreviewAttachment?: (attachment: AttachmentListItem) => void;
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  onOpenAllArtifacts: (outputs: WorkspaceOutputRecordPayload[]) => void;
-  collapsedTraceByStepId: Record<string, boolean>;
-  onToggleTraceStep: (stepId: string) => void;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-  memoryProposalAction: {
-    proposalId: string;
-    action: "accept" | "dismiss";
-  } | null;
-  editingMemoryProposalId: string | null;
-  memoryProposalDrafts: Record<string, string>;
-  onEditMemoryProposal: (message: Message, proposalId: string) => void;
-  onMemoryProposalDraftChange: (proposalId: string, value: string) => void;
-  onAcceptMemoryProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-  onDismissMemoryProposal: (
-    proposal: MemoryUpdateProposalRecordPayload,
-  ) => void;
-  assistantFooterAccessoryMessageId?: string | null;
-  assistantFooterAccessory?: ReactNode;
-  getMessageWrapperClassName?: (message: Message) => string | undefined;
-  liveAssistantTurn?: {
-    text: string;
-    tone?: ChatMessage["tone"];
-    segments: ChatAssistantSegment[];
-    executionItems: ChatExecutionTimelineItem[];
-    status?: string;
-    statusAccessory?: ReactNode;
-    footerAccessory?: ReactNode;
-  } | null;
-}) {
-  return (
-    <>
-      {messages.map((message, index) => {
-        const wrapperClassName = getMessageWrapperClassName?.(message)?.trim();
-        const next = messages[index + 1];
-        const isLastInAssistantGroup =
-          message.role === "assistant" &&
-          (!next || next.role === "user") &&
-          !liveAssistantTurn;
-        const turn =
-          message.role === "user" ? (
-            <UserTurn
-              text={message.text}
-              createdAt={message.createdAt}
-              attachments={message.attachments ?? []}
-              onPreviewAttachment={onPreviewAttachment}
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-            />
-          ) : (
-            <AssistantTurn
-              label={assistantLabel}
-              mode={assistantMode}
-              showExecutionInternals={showExecutionInternals}
-              fitToContent={assistantFitToContent}
-              text={message.text}
-              tone={message.tone ?? "default"}
-              segments={message.segments ?? []}
-              executionItems={message.executionItems ?? []}
-              memoryProposals={message.memoryProposals ?? []}
-              outputs={message.outputs ?? []}
-              memoryProposalAction={memoryProposalAction}
-              editingMemoryProposalId={editingMemoryProposalId}
-              memoryProposalDrafts={memoryProposalDrafts}
-              onEditMemoryProposal={(proposalId) =>
-                onEditMemoryProposal(message, proposalId)
-              }
-              onMemoryProposalDraftChange={onMemoryProposalDraftChange}
-              onAcceptMemoryProposal={onAcceptMemoryProposal}
-              onDismissMemoryProposal={onDismissMemoryProposal}
-              onOpenOutput={onOpenOutput}
-              onOpenAllArtifacts={onOpenAllArtifacts}
-              collapsedTraceByStepId={collapsedTraceByStepId}
-              onToggleTraceStep={onToggleTraceStep}
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-              showAvatar={isLastInAssistantGroup}
-              workspaceId={workspaceId ?? null}
-              createdAt={message.createdAt}
-              footerAccessory={
-                message.id === assistantFooterAccessoryMessageId
-                  ? assistantFooterAccessory
-                  : null
-              }
-            />
-          );
-
-        if (wrapperClassName) {
-          return (
-            <div key={message.id} className={wrapperClassName}>
-              {turn}
-            </div>
-          );
-        }
-        return <Fragment key={message.id}>{turn}</Fragment>;
-      })}
-
-      {liveAssistantTurn ? (
-        <AssistantTurn
-          label={assistantLabel}
-          mode={assistantMode}
-          showExecutionInternals={showExecutionInternals}
-          fitToContent={assistantFitToContent}
-          text={liveAssistantTurn.text}
-          tone={liveAssistantTurn.tone ?? "default"}
-          segments={liveAssistantTurn.segments}
-          executionItems={liveAssistantTurn.executionItems}
-          memoryProposals={[]}
-          outputs={[]}
-          memoryProposalAction={memoryProposalAction}
-          editingMemoryProposalId={editingMemoryProposalId}
-          memoryProposalDrafts={memoryProposalDrafts}
-          onEditMemoryProposal={() => undefined}
-          onMemoryProposalDraftChange={onMemoryProposalDraftChange}
-          onAcceptMemoryProposal={onAcceptMemoryProposal}
-          onDismissMemoryProposal={onDismissMemoryProposal}
-          onOpenOutput={onOpenOutput}
-          onOpenAllArtifacts={onOpenAllArtifacts}
-          collapsedTraceByStepId={collapsedTraceByStepId}
-          onToggleTraceStep={onToggleTraceStep}
-          onLinkClick={onLinkClick}
-          onLocalLinkClick={onLocalLinkClick}
-          showAvatar
-          workspaceId={workspaceId ?? null}
-          live
-          statusAccessory={liveAssistantTurn.statusAccessory ?? null}
-          status={liveAssistantTurn.status ?? ""}
-          footerAccessory={liveAssistantTurn.footerAccessory ?? null}
-        />
-      ) : null}
-    </>
   );
 }
